Fixes ano#501

+++ ebay/sql/sql-install.php
@@ -107,7 +107,7 @@ $sql[] = 'CREATE TABLE IF NOT EXISTS `'._DB_PREFIX_.'ebay_shipping` (
 		  `id_ebay_shipping` int(11) NOT NULL AUTO_INCREMENT,
 		  `ebay_carrier` varchar(256) NOT NULL,
 		  `ps_carrier` int(11) NOT NULL,
-		  `extra_fee` int(11) NOT NULL,
+		  `extra_fee` float(8,2) NOT NULL,
 		  `international` int(4) NOT NULL, 
 		  PRIMARY KEY (`id_ebay_shipping`)
 	) ENGINE='._MYSQL_ENGINE_.'  DEFAULT CHARSET=utf8';
